LC-16709: Add comprehensive reference documentation
- Create docs/language.md with expression language reference and examples
- Create docs/functions.md with 43 built-in functions sorted alphabetically
- Create docs/api.md with TypeScript API reference and error hierarchy
- Create docs/benchmark.md with CLI usage guide and library comparison
- Streamline README.md to link to documentation
- Move internal helpers (invertedSlice, invertedFilter, replaceArray) to access-internal.ts
- Move invertedSlice tests to test/access-internal.test.ts
- Fix array signatures in functions.md to use TypeScript syntax (type[])
- Add comprehensive error documentation and conditions for all functions
